Reorganization of routing classes
